requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
joblib==1.3.2
scikit-learn==1.3.2
sqlalchemy==2.0.23
python-multipart==0.0.6
orjson==3.9.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0